requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
anthropic==0.81.0
instructor==1.14.5
ollama==0.6.1
sqlite-vec==0.1.6
tenacity==9.1.4
pydantic==2.12.5
pydantic-settings==2.13.0
python-dotenv==1.2.1
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
google-api-python-client==2.190.0
numpy==1.26.4
mcp==1.26.0
pyyaml==6.0.3
httpx==0.28.1
starlette==0.52.1
# Document intelligence
langextract==1.1.1
pymupdf==1.27.1
google-genai==1.63.0
# Push notifications
aioapns==4.0
# Daemon scheduling
apscheduler==4.0.0a6
# HTTP client for daemon async requests
aiohttp==3.13.3
# Credential encryption for users
cryptography==46.0.5
# Error tracking
sentry-sdk[fastapi]==2.53.0
# School calendar ICS parsing
icalendar==7.0.0
recurring-ical-events==3.8.1
# Channel adapters
python-telegram-bot==22.6
discord.py==2.6.4
